Add files via upload
The program is for predict the closing price by the past trading days with machine learning. Users have to enter the stock code, starting and ending dates, and the number of trading days they would like to take as a reference to predict the closing price. The prediction result is compared with the actual closing price in the recent trading days. It is for testing whether the function of prediction is accurate. If the result is close to the actual performance, further prediction can be performed to forecast the future stock price.
**That is not for stock price forecasting
**It is just a practice for machine-learning
**It should not be taken as professional investment advice